+++ soa_sortMatrices.m
@@ -1,4 +1,5 @@
-% Objective: Sort the contents of the matrices.
+% Objective: Row-wise sorting of matrix dataMatrix according to indexMatrix.
+%            Used to sort quantities according to the difference tauO - tauA.
 %
 % Copyright: Roberto Legaspi and Taro Toyoizumi
 %            Lab for Neural Computation and Adaptation
@@ -10,11 +11,17 @@
 %
 % SPDX-License-Identifier: CC-BY-4.0
 
-function sortedMatrix = ...
-    soa_sortMatrices(vectMatrix, sortIndx1, sortIndx2, sortIndx3)
+function sortedMatrix = soa_sortMatrices(dataMatrix, indexMatrix)
 
-sortedMatrix = nan(size(vectMatrix));
-sortedMatrix(1, :) = vectMatrix(1, sortIndx1);
-sortedMatrix(2, :) = vectMatrix(2, sortIndx2);
-sortedMatrix(3, :) = vectMatrix(3, sortIndx3);
+size_data = size(dataMatrix);
+size_indx = size(indexMatrix);
+
+assert(all(size_data == size_indx), 'Matrices must have same dimensions.')
+assert(numel(size_data) == 2, 'Matrices must be two-dimensional.')
+
+sortedMatrix = nan(size_data);
+
+for row_idx = 1:size_data(1)
+    sortedMatrix(row_idx, :) = dataMatrix(row_idx, indexMatrix(row_idx, :));
+end
 end